Refactor: models package created, adapters package moved, removal of createNewTodo
